Actualizacion proyecto

+++ aplicacion/Principal.java
@@ -3,6 +3,6 @@ package aplicacion;
 import mates.*;
 public class Principal{
     public static void main(String[] args){
-           System.out.println("El número PI es " + Matematicas.generarNumeroPi(Integer.parseInt(args[0])));
+           System.out.println("El número PI es " + Matematicas.generarNumeroPi(Long.parseLong(args[0])));
         }
 }
